+++ domain/review-domain-rdb/src/main/java/woowa/team4/bff/review/service/ReviewService.java
@@ -3,10 +3,7 @@ package woowa.team4.bff.review.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.event.TransactionPhase;
-import org.springframework.transaction.event.TransactionalEventListener;
 import woowa.team4.bff.event.review.ReviewCreateEvent;
 import woowa.team4.bff.menu.item.repository.MenuRepository;
 import woowa.team4.bff.publisher.EventPublisher;
@@ -14,14 +11,10 @@ import woowa.team4.bff.restaurant.repository.RestaurantFinder;
 import woowa.team4.bff.review.command.ReviewCreateCommand;
 import woowa.team4.bff.review.domain.Review;
 import woowa.team4.bff.review.domain.ReviewMenu;
-import woowa.team4.bff.review.domain.ReviewStatistics;
-import woowa.team4.bff.review.entity.ReviewStatisticsEntity;
 import woowa.team4.bff.review.repository.ReviewMenuRepository;
 import woowa.team4.bff.review.repository.ReviewRepository;
-import woowa.team4.bff.review.repository.ReviewStatisticsRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -30,7 +23,6 @@ public class ReviewService {
 
     private final ReviewRepository reviewRepository;
     private final ReviewMenuRepository reviewMenuRepository;
-    private final ReviewStatisticsRepository reviewStatisticsRepository;
     private final RestaurantFinder restaurantFinder;
     private final MenuRepository menuRepository;
     private final EventPublisher eventPublisher;
@@ -61,7 +53,7 @@ public class ReviewService {
         eventPublisher.publish(new ReviewCreateEvent(restaurantId, savedReview.getRating()));
 
         log.info("[ReviewMenus] : {}", savedReviewMenus);
-        log.info("[Review] : {}", savedReview.toString());
+        log.info("[Review] : {}", savedReview);
 
         return ReviewCreateResult.builder()
                 .reviewUuid(savedReview.getReviewUuId())
@@ -69,33 +61,4 @@ public class ReviewService {
                         .map(ReviewMenu::getReviewMenuUuid).toList())
                 .build();
     }
-
-    @Async
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public void updateReviewStatistics(ReviewCreateEvent event) {
-
-        Optional<ReviewStatistics> optionalReviewStatistics = reviewStatisticsRepository.findByRestaurantId(event.restaurantId());
-        if(optionalReviewStatistics.isEmpty()) {
-            ReviewStatisticsEntity reviewStatisticsEntity = ReviewStatisticsEntity.builder()
-                    .restaurantId(event.restaurantId())
-                    .reviewCount(1L)
-                    .averageRating(event.rating())
-                    .build();
-            ReviewStatistics save = reviewStatisticsRepository.save(reviewStatisticsRepository.toDomain(reviewStatisticsEntity));
-            log.info("[create ReviewStatistics] : {}", save);
-            return;
-        }
-
-        ReviewStatistics reviewStatistics = optionalReviewStatistics.get();
-        ReviewStatistics savedReviewStatistic = reviewStatisticsRepository.save(updateReviewStatistics(reviewStatistics, event));
-    }
-
-    public ReviewStatistics updateReviewStatistics(ReviewStatistics reviewStatistics, ReviewCreateEvent event) {
-        double totalRating = reviewStatistics.getReviewCount() * reviewStatistics.getAverageRating() + event.rating();
-        long reviews = reviewStatistics.getReviewCount() + 1;
-        reviewStatistics.setReviewCount(reviews);
-        reviewStatistics.setAverageRating(totalRating / reviews);
-        return reviewStatistics;
-    }
-
 }
